+++ src/schema-validator.js
@@ -14,68 +14,73 @@ import { dirname, join } from 'path';
 import { fileURLToPath } from 'url';
 import { readFile, readdir } from 'fs/promises';
 
-const SCHEMA_INGESTION_REQUEST = 'ingestion-request';
+export const SCHEMA_INGESTION_REQUEST = 'ingestion-request';
 
 /**
  * Loads schemas from this project in the <code>schemas</code> directory and supports validating
  * objects against the schemas.
  */
 export class SchemaValidator {
-  #loaded = false;
-
   #schemas = {};
 
-  async #getSchema(name) {
-    if (!this.#loaded) {
-      const schemasDir = join(
-        dirname(fileURLToPath(import.meta.url)),
-        'schemas',
-      );
-      const schemas = await readdir(schemasDir);
-      await Promise.all(
-        schemas
-          .filter((fileName) => fileName.endsWith('.json'))
-          .map(async (fileName) => {
-            const schemaName = fileName.slice(0, -5);
-            const schemaBuf = await readFile(join(schemasDir, fileName));
-            this.#schemas[schemaName] = JSON.parse(schemaBuf.toString());
-          }),
-      );
-      this.#loaded = true;
+  async loadSchemas(dir) {
+    const schemasDir = dir || join(dirname(fileURLToPath(import.meta.url)), 'schemas');
+    const schemaFiles = await readdir(schemasDir);
+    await Promise.all(
+      schemaFiles.filter((fileName) => fileName.endsWith('.json'))
+        .map(async (fileName) => {
+          this.#schemas[fileName.slice(0, -5)] = JSON.parse(
+            (await readFile(join(schemasDir, fileName))).toString(),
+          );
+        }),
+    );
+  }
+
+  async getSchema(name, dir) {
+    if (!Object.keys(this.#schemas).length) {
+      await this.loadSchemas(dir);
     }
     return this.#schemas[name];
   }
 
   /**
-   * Validates the <code>ingestionRequest</code> object against the Ingestion Request schema
+   * Validates an object against the schema specified by <code>schemaName</code>
    * as specified in https://wiki.corp.adobe.com/display/WEM/Ingestor+API+Contract
+   *
+   * Throws <code>Error</code> if request does not match
+   *
    * @see https://wiki.corp.adobe.com/display/WEM/Ingestor+API+Contract?
-   * @param {any} ingestionRequest
+   * @param {any} obj
+   * @param {string} schemaName
    * @param {Array<string>} additionalRequiredData
    */
-  async validateIngestionRequest(ingestionRequest, additionalRequiredData) {
-    const schema = await this.#getSchema(SCHEMA_INGESTION_REQUEST);
-    const result = validate(ingestionRequest, schema, {
+  async validateObject(obj, schemaName, additionalRequiredData) {
+    const schema = await this.getSchema(schemaName);
+    const result = validate(obj, schema, {
       allowUnknownAttributes: false,
     });
     if (result.errors?.length > 0) {
       throw new Error(
-        `Failed to validate schema, errors: \n- ${result.errors
+        `Failed to validate schema "${schemaName}", errors: \n- ${result.errors
           .map((e) => e.message)
           .join('\n- ')}`,
       );
     }
     if (additionalRequiredData) {
       const missing = additionalRequiredData.filter(
-        (key) => !(key in ingestionRequest.data),
+        (key) => !(key in obj.data),
       );
       if (missing.length > 0) {
         throw new Error(
-          `Failed to validate schema, missing data fields: [${missing.join(
+          `Failed to validate schema "${schemaName}", missing data fields: [${missing.join(
             ', ',
           )}]`,
         );
       }
     }
+  }
+
+  async validateIngestionRequest(request, additionalRequiredData) {
+    return this.validateRequest(request, SCHEMA_INGESTION_REQUEST, additionalRequiredData);
   }
 }