+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.7.0](https://github.com/adobe/content-lake-commons/compare/v1.6.1...v1.7.0) (2023-05-25)
+
+
+### Features
+
+* Add ingestor internal schemas ([#30](https://github.com/adobe/content-lake-commons/issues/30)) ([c6d342e](https://github.com/adobe/content-lake-commons/commit/c6d342eca9dc41c9b4eebae833a43665b6042f98))
+
+## [1.6.1](https://github.com/adobe/content-lake-commons/compare/v1.6.0...v1.6.1) (2023-05-25)
+
+
+### Bug Fixes
+
+* adding delete to the search mock ([#29](https://github.com/adobe/content-lake-commons/issues/29)) ([72a55b8](https://github.com/adobe/content-lake-commons/commit/72a55b85048f15183908ed141660725371092d56))
+
 # [1.6.0](https://github.com/adobe/content-lake-commons/compare/v1.5.1...v1.6.0) (2023-05-23)